promote output module to directory
